refactor: move filters into band panel and use equal panel heights
